refactor(weapon): 拆分 WeaponOverview.vue 为弹窗组件与 composables
将 2178 行的 WeaponOverview.vue 按职责拆分：

组件拆分
- WeaponOverview.vue：1630 → 707 行，仅保留网格区、连线系统与编排逻辑
- WeaponDetailDialog.vue（新增）：详情弹窗 + 删除确认弹窗，编辑状态/防抖自动保存/swapMatrix 全迁入，对外仅暴露 v-model:weapon-id
- OverlapDetectionDialog.vue（新增）：重合检测弹窗，零 props

composables 抽取
- useWeaponStats / useCustomStats / useOverlapDetection / useWeaponConnectionLines
- 状态 ref 提升为模块级单例，保证父组件与弹窗共享同一份数据
- 修复拆分引入的重合对比失效（matrixEntryByWeaponId 传空 Map）与 stat_types 语义改变
