+++ frontend/src/components/WeaponOverview.vue
@@ -235,372 +235,67 @@
     </v-expansion-panel-text>
   </v-expansion-panel>
 
-  <!-- 武器详情弹窗 -->
-  <v-dialog v-model="detailDialog" max-width="680">
-    <v-card v-if="detailWeaponId">
-      <v-card-item>
-        <template #prepend>
-          <custom-stat-icon v-if="isCustomEntry(detailWeaponId) || isNewCustom" class="weapon-icon-detail" hide-name :name="customEntryName || '新基质'" :skill-stat-id="customEditSkill" />
-          <item-icon v-else class="weapon-icon-detail" :item-id="detailWeaponId" />
-        </template>
-        <v-card-title>
-          <v-text-field
-            v-if="isCustomEntry(detailWeaponId) || isNewCustom"
-            v-model="customEntryName"
-            density="compact"
-            hide-details
-            placeholder="自定义基质名称"
-            variant="underlined"
-          />
-          <template v-else>
-            {{ weaponsMap.get(detailWeaponId)?.name || detailWeaponId }}
-          </template>
-        </v-card-title>
-        <template #append>
-          <v-btn icon="mdi-close" variant="text" @click="detailDialog = false" />
-        </template>
-      </v-card-item>
-      <v-divider />
-      <v-card-text>
-        <!-- 基质属性 -->
-        <div class="mb-4">
-          <div class="text-subtitle-2 mb-2">基质属性</div>
-          <v-row dense>
-            <v-col cols="12" sm="4">
-              <v-select
-                :clearable="isCustomEntry(detailWeaponId) || isNewCustom"
-                density="compact"
-                :disabled="!(isCustomEntry(detailWeaponId) || isNewCustom)"
-                hide-details
-                :items="allAttributeStats.map(id => ({ title: getGemTagName(id), value: id }))"
-                label="基础属性"
-                :model-value="isCustomEntry(detailWeaponId) || isNewCustom ? customEditAttribute : weaponsMap.get(detailWeaponId)?.attributeStatId ?? null"
-                variant="outlined"
-                @update:model-value="isCustomEntry(detailWeaponId) || isNewCustom ? customEditAttribute = $event : undefined"
-              />
-            </v-col>
-            <v-col cols="12" sm="4">
-              <v-select
-                :clearable="isCustomEntry(detailWeaponId) || isNewCustom"
-                density="compact"
-                :disabled="!(isCustomEntry(detailWeaponId) || isNewCustom)"
-                hide-details
-                :items="allSecondaryStats.map(id => ({ title: getGemTagName(id), value: id }))"
-                label="附加属性"
-                :model-value="isCustomEntry(detailWeaponId) || isNewCustom ? customEditSecondary : weaponsMap.get(detailWeaponId)?.secondaryStatId ?? null"
-                variant="outlined"
-                @update:model-value="isCustomEntry(detailWeaponId) || isNewCustom ? customEditSecondary = $event : undefined"
-              />
-            </v-col>
-            <v-col cols="12" sm="4">
-              <v-select
-                :clearable="isCustomEntry(detailWeaponId) || isNewCustom"
-                density="compact"
-                :disabled="!(isCustomEntry(detailWeaponId) || isNewCustom)"
-                hide-details
-                :items="allSkillStats.map(id => ({ title: getGemTagName(id), value: id }))"
-                label="技能属性"
-                :model-value="isCustomEntry(detailWeaponId) || isNewCustom ? customEditSkill : weaponsMap.get(detailWeaponId)?.skillStatId ?? null"
-                variant="outlined"
-                @update:model-value="isCustomEntry(detailWeaponId) || isNewCustom ? customEditSkill = $event : undefined"
-              />
-            </v-col>
-          </v-row>
-        </div>
-
-        <!-- 基质等级 -->
-        <div class="mb-4 detail-level-outer">
-          <div class="text-subtitle-2 mb-2">当前基质等级</div>
-          <div class="detail-level-wrapper">
-            <div class="detail-level-section">
-              <div class="detail-attr-control detail-attr-control--primary">
-                <span class="detail-attr-label">基础属性</span>
-                <div class="detail-attr-pips">
-                  <span
-                    v-for="level in affixLevelItems"
-                    :key="`d-a1-${level}`"
-                    class="detail-pip"
-                    :class="{
-                      active: level <= detailAffix1,
-                      'detail-pip--max': detailAffix1 === 6,
-                    }"
-                    @click="detailAffix1 = level"
-                  />
-                </div>
-                <span class="detail-attr-value" :class="{ 'detail-attr-value--full': detailAffix1 === 6 }">
-                  +{{ detailAffix1 }} / 6
-                </span>
-              </div>
-
-              <div class="detail-attr-control detail-attr-control--teal">
-                <span class="detail-attr-label">附加属性</span>
-                <div class="detail-attr-pips">
-                  <span
-                    v-for="level in affixLevelItems"
-                    :key="`d-a2-${level}`"
-                    class="detail-pip"
-                    :class="{
-                      active: level <= detailAffix2,
-                      'detail-pip--max': detailAffix2 === 6,
-                    }"
-                    @click="detailAffix2 = level"
-                  />
-                </div>
-                <span class="detail-attr-value" :class="{ 'detail-attr-value--full': detailAffix2 === 6 }">
-                  +{{ detailAffix2 }} / 6
-                </span>
-              </div>
-
-              <div class="detail-attr-control detail-attr-control--indigo">
-                <span class="detail-attr-label">技能属性</span>
-                <div class="detail-attr-pips detail-attr-pips--skill">
-                  <span
-                    v-for="level in skillLevelItems"
-                    :key="`d-a3-${level}`"
-                    class="detail-pip"
-                    :class="{
-                      active: level <= detailAffix3,
-                      'detail-pip--max': detailAffix3 === 3,
-                    }"
-                    @click="detailAffix3 = level"
-                  />
-                </div>
-                <span class="detail-attr-value" :class="{ 'detail-attr-value--full': detailAffix3 === 3 }">
-                  +{{ detailAffix3 }} / 3
-                </span>
-              </div>
-            </div>
-
-            <!-- 未拥有斜向胶带遮罩（贯穿三个属性，可点击切换拥有） -->
-            <transition name="tape-peel">
-              <div
-                v-if="!isDetailOwned"
-                class="not-owned-tape-detail"
-                @click="toggleDetailOwnership"
-              >
-                <span class="not-owned-tape-detail-text">» 未拥有 » NOT OWNED » 点击切换为拥有 » CLICK TO TOGGLE » </span>
-              </div>
-            </transition>
-          </div>
-
-          <!-- 未拥有遮罩：覆盖等级区域，阻止交互 -->
-          <div v-if="!isDetailOwned" class="not-owned-block-overlay" @click="toggleDetailOwnership" />
-
-          <!-- 拥有状态切换 -->
-          <div class="mt-2">
-            <v-chip
-              :color="isDetailOwned ? 'success' : 'grey'"
-              :prepend-icon="isDetailOwned ? 'mdi-check-circle' : 'mdi-close-circle'"
-              size="small"
-              variant="tonal"
-              @click="toggleDetailOwnership"
-            >
-              {{ isDetailOwned ? '已拥有 · 点击可切换为 未拥有' : '未拥有 · 点击可切换为 已拥有' }}
-            </v-chip>
-          </div>
-        </div>
-
-        <!-- 优先级设置 -->
-        <div class="mb-4">
-          <div class="text-subtitle-2 mb-1">基质匹配优先级</div>
-          <div class="d-flex flex-wrap align-center ga-2 mb-2">
-            <v-chip
-              v-for="p in [1, 2, 3, 4, 5, 6, 7, 8, 9]"
-              :key="p"
-              :color="detailPriority === p ? 'primary' : undefined"
-              size="small"
-              :variant="detailPriority === p ? 'flat' : 'outlined'"
-              @click="detailPriority = p"
-            >
-              {{ p }}
-            </v-chip>
-          </div>
-          <div class="text-caption text-medium-emphasis" style="line-height: 1.6">
-            当扫描到一个无暇基质同时匹配多把武器时，系统会按优先级将该基质分配给优先级最高的武器。<br />
-            默认使用武器稀有度作为优先级（6★=6, 5★=5, 4★=4, 3★=3）。<br />
-            手动设置 1-9 可覆盖默认值，数值越大越优先。<br />
-            已满级（6/6/3）的武器会被自动跳过。
-          </div>
-        </div>
-
-        <!-- 同类武器（仅非自定义且非新建时显示） -->
-        <div v-if="!isCustomEntry(detailWeaponId) && !isNewCustom && getSameStatWeapons(detailWeaponId).length > 0">
-          <div class="text-subtitle-2 mb-2">同类属性武器</div>
-          <div class="d-flex flex-column ga-2">
-            <v-card
-              v-for="sameId in getSameStatWeapons(detailWeaponId)"
-              :key="sameId"
-              class="pa-2"
-              variant="outlined"
-            >
-              <div class="d-flex align-center justify-space-between">
-                <div class="d-flex align-center ga-2">
-                  <item-icon class="weapon-icon-same" :item-id="sameId" />
-                  <div>
-                    <div class="font-weight-bold text-body-2">
-                      {{ weaponsMap.get(sameId)?.name || sameId }}
-                    </div>
-                    <div class="text-caption text-medium-emphasis">
-                      {{ getMatrixLevelText(sameId) }}
-                      <span class="ml-1">优先级: {{ getWeaponPriority(sameId) }}</span>
-                    </div>
-                  </div>
-                </div>
-                <v-btn
-                  color="primary"
-                  size="small"
-                  variant="tonal"
-                  @click="swapMatrix(detailWeaponId!, sameId)"
-                >
-                  交换
-                </v-btn>
-              </div>
-            </v-card>
-          </div>
-        </div>
-        <div v-else-if="!isCustomEntry(detailWeaponId) && !isNewCustom" class="text-medium-emphasis text-caption">
-          没有其他武器与此武器共享相同属性组合。
-        </div>
-      </v-card-text>
-
-      <!-- 操作按钮 -->
-      <v-divider />
-      <v-card-actions>
-        <!-- 自定义基质（新建+编辑）：保存 + 删除 -->
-        <template v-if="isCustomEntry(detailWeaponId) || isNewCustom">
-          <v-btn
-            v-if="!isNewCustom"
-            color="error"
-            prepend-icon="mdi-delete"
-            variant="text"
-            @click="promptDeleteCustomEntry"
-            @mousedown="pendingCustomDelete = true"
-          >
-            删除此自定义基质·无法恢复
-          </v-btn>
-          <v-spacer />
-          <v-btn
-            color="primary"
-            prepend-icon="mdi-content-save"
-            variant="flat"
-            @click="saveCustomEntry"
-          >
-            保存
-          </v-btn>
-        </template>
-        <!-- 非自定义基质：清空 -->
-        <template v-else>
-          <v-btn
-            v-if="isWeaponOwned(detailWeaponId)"
-            color="error"
-            prepend-icon="mdi-delete-outline"
-            variant="text"
-            @click="removeNonCustomEntry(detailWeaponId!)"
-          >
-            清空此基质的保存数据
-          </v-btn>
-          <v-spacer />
-        </template>
-      </v-card-actions>
-    </v-card>
-  </v-dialog>
+  <!-- 武器详情弹窗（含删除确认弹窗） -->
+  <weapon-detail-dialog v-model:weapon-id="detailWeaponId" />
 
   <!-- 自定义基质重合检测弹窗 -->
-  <v-dialog v-model="overlapDialog" max-width="700" persistent>
-    <v-card>
-      <v-card-title class="d-flex align-center">
-        <v-icon class="mr-2" color="warning">mdi-alert-circle-outline</v-icon>
-        自定义基质重合检测
-      </v-card-title>
-      <v-card-text>
-        <v-alert class="mb-4" type="info" variant="tonal">
-          以下自定义基质与内置武器词条重复，可选择移除或切换。
-        </v-alert>
-        <div
-          v-for="(item, idx) in overlapItems"
-          :key="idx"
-          class="d-flex align-center mb-3 pa-3 border rounded flex-wrap"
-          style="gap: 8px"
-        >
-          <v-tooltip location="top" open-delay="0">
-            <template #activator="{ props: tooltipProps }">
-              <div v-bind="tooltipProps" class="overlap-icon-wrapper" :class="{ 'weapon-not-owned': !isWeaponOwned(item.customWeaponId) }">
-                <custom-stat-icon hide-name :name="item.customName" :skill-stat-id="customStats[item.customIndex]?.skill ?? null" small/>
-              </div>
-            </template>
-            <span>{{ getOverlapTooltipText(item.customWeaponId) }}</span>
-          </v-tooltip>
-          <span class="font-weight-bold">{{ item.customName }}</span>
-          <v-icon
-            :color="getOverlapCompareResult(item) > 0 ? 'success' : getOverlapCompareResult(item) < 0 ? 'error' : 'grey'"
-            size="small"
-          >
-            {{ getOverlapCompareResult(item) > 0 ? 'mdi-arrow-right-bold' : getOverlapCompareResult(item) < 0 ? 'mdi-arrow-left-bold' : 'mdi-arrow-left-right' }}
-          </v-icon>
-          <v-tooltip location="top" open-delay="0">
-            <template #activator="{ props: tooltipProps }">
-              <div v-bind="tooltipProps" class="overlap-icon-wrapper" :class="{ 'weapon-not-owned': !isWeaponOwned(item.matchedWeaponId) }">
-                <item-icon class="weapon-icon-overlap" :item-id="item.matchedWeaponId" />
-              </div>
-            </template>
-            <span>{{ getOverlapTooltipText(item.matchedWeaponId) }}</span>
-          </v-tooltip>
-          <span class="font-weight-bold">{{ item.matchedWeaponName }}</span>
-          <v-spacer />
-          <v-chip-group v-model="item.action" mandatory>
-            <v-chip color="error" filter size="small" value="delete" variant="outlined">移除</v-chip>
-            <v-chip color="primary" filter size="small" value="switch" variant="outlined">切换</v-chip>
-            <v-chip filter size="small" value="ignore" variant="outlined">本次忽略</v-chip>
-            <v-chip filter size="small" value="suppress" variant="outlined">不再提示</v-chip>
-          </v-chip-group>
-        </div>
-      </v-card-text>
-      <v-card-actions>
-        <v-btn color="secondary" variant="outlined" @click="autoSelectOverlap">一键选择</v-btn>
-        <v-spacer />
-        <v-btn @click="overlapDialog = false">取消</v-btn>
-        <v-btn color="primary" @click="confirmOverlapActions">确认</v-btn>
-      </v-card-actions>
-    </v-card>
-  </v-dialog>
-
-  <!-- 删除自定义基质确认弹窗 -->
-  <v-dialog v-model="deleteCustomConfirm" max-width="460">
-    <v-card>
-      <v-card-title class="text-error">删除自定义基质</v-card-title>
-      <v-card-text>
-        确定要删除自定义基质「{{ deleteCustomName }}」吗？将从设置中移除该基质定义，并从当前账号的基质数据中移除对应条目，此操作不可撤销。
-      </v-card-text>
-      <v-card-actions>
-        <v-spacer />
-        <v-btn @click="cancelDeleteCustomEntry">取消</v-btn>
-        <v-btn color="error" @click="confirmDeleteCustomEntry">删除</v-btn>
-      </v-card-actions>
-    </v-card>
-  </v-dialog>
+  <overlap-detection-dialog />
 </template>
 
 <script lang="ts" setup>
-import { computed, onMounted, onUnmounted, ref, watch } from 'vue'
+import { computed, onMounted, ref, watch } from 'vue'
 import CustomStatIcon from '@/components/CustomStatIcon.vue'
 import ItemIcon from '@/components/ItemIcon.vue'
-import { type TreasureMatrixEntry, useProfiles } from '@/composables/useProfiles'
+import OverlapDetectionDialog from '@/components/OverlapDetectionDialog.vue'
+import WeaponDetailDialog from '@/components/WeaponDetailDialog.vue'
+import { useCustomStats } from '@/composables/useCustomStats'
+import { useOverlapDetection } from '@/composables/useOverlapDetection'
+import { useProfiles } from '@/composables/useProfiles'
 import { useRarityFilters } from '@/composables/useRarityFilters'
+import { useWeaponConnectionLines } from '@/composables/useWeaponConnectionLines'
+import { useWeaponStats } from '@/composables/useWeaponStats'
 import { useStaticData } from '@/utils/gameData/staticData'
-import { type CustomStat, getGemTagName } from '@/utils/gameData/weapon'
+import { getGemTagName } from '@/utils/gameData/weapon'
 
-const { weaponTypes, weaponsMap, essencesMap, matrixIcons } = useStaticData()
+const { weaponTypes, weaponsMap, matrixIcons } = useStaticData()
 const {
   activeProfile,
-  treasureMatrix,
-  addTreasureMatrixEntry,
-  removeTreasureMatrixEntry,
-  updateTreasureMatrix,
   updateSwitchDisplayMode,
   updateMatrixBadgeDisplayMode,
-  updateWeaponPriority,
 } = useProfiles()
 const { selectedRarities } = useRarityFilters()
+
+// 使用 composables
+const {
+  matrixEntryByWeaponId,
+  ownedWeaponIds,
+  isWeaponOwned,
+  isWeaponMaxed,
+  isCustomEntry,
+  getSameStatWeapons,
+  isSwitchable,
+  isSwitchTargetMaxed,
+} = useWeaponStats()
+
+const {
+  customStats,
+  customMatrixEntries,
+  fetchCustomStats,
+  toggleWeaponOwnership,
+} = useCustomStats()
+
+const {
+  checkCustomOverlap: checkCustomOverlapBase,
+} = useOverlapDetection()
+
+const {
+  containerRef,
+  connectionLines,
+  hoveredWeaponId,
+  handleWeaponMouseEnter: handleWeaponMouseEnterBase,
+  handleWeaponMouseLeave,
+  setupResizeListener,
+} = useWeaponConnectionLines()
 
 // 可切换提示显示模式：'chip'=大号提示(默认), 'dot'=小橙点, 'off'=关闭
 type SwitchDisplayMode = 'chip' | 'dot' | 'off'
@@ -672,808 +367,40 @@ function getWeaponSkillIcon(weaponId: string): string | null {
   return matrixIcons.value.skills[weapon.skillStatId] || null
 }
 
-// --- 属性选项列表 ---
-const allAttributeStats = computed(() =>
-  Array.from(essencesMap.value.values())
-    .filter((e) => e.type === 'ATTRIBUTE')
-    .map((e) => e.id),
-)
-const allSecondaryStats = computed(() =>
-  Array.from(essencesMap.value.values())
-    .filter((e) => e.type === 'SECONDARY')
-    .map((e) => e.id),
-)
-const allSkillStats = computed(() =>
-  Array.from(essencesMap.value.values())
-    .filter((e) => e.type === 'SKILL')
-    .map((e) => e.id),
-)
-
-// --- 等级选项 ---
-const affixLevelItems = [1, 2, 3, 4, 5, 6]
-const skillLevelItems = [1, 2, 3]
-
-// 武器详情弹窗
-const detailDialog = ref(false)
+// 武器详情弹窗（null=关闭，string=打开编辑该武器）
 const detailWeaponId = ref<string | null>(null)
 
-// --- 武器连线系统 ---
-
-/** 容器引用 */
-const containerRef = ref<HTMLElement | null>(null)
-
-/** 当前悬停的武器 ID */
-const hoveredWeaponId = ref<string | null>(null)
-
-/** 连线数据 */
-interface ConnectionLine {
-  style: {
-    left: string
-    top: string
-    width: string
-    transform: string
-    opacity: number
-  }
-}
-
-const connectionLines = ref<ConnectionLine[]>([])
-
-/** 获取武器图标元素位置 */
-function getWeaponElementPosition(weaponId: string): { x: number; y: number } | null {
-  const container = containerRef.value
-  if (!container) return null
-
-  const element = container.querySelector(`[data-weapon-id="${weaponId}"]`)
-  if (!element) return null
-
-  const containerRect = container.getBoundingClientRect()
-  const elementRect = element.getBoundingClientRect()
-
-  return {
-    x: elementRect.left - containerRect.left + elementRect.width / 2,
-    y: elementRect.top - containerRect.top + elementRect.height / 2,
-  }
-}
-
-/** 计算两点之间的距离和角度 */
-function calculateLine(
-  startX: number,
-  startY: number,
-  endX: number,
-  endY: number,
-): { length: number; angle: number } {
-  const dx = endX - startX
-  const dy = endY - startY
-  const length = Math.hypot(dx, dy)
-  const angle = Math.atan2(dy, dx) * (180 / Math.PI)
-  return { length, angle }
-}
-
 /** 更新连线 */
-function updateConnectionLines() {
-  if (!hoveredWeaponId.value) {
-    connectionLines.value = []
-    return
-  }
-
-  const startPos = getWeaponElementPosition(hoveredWeaponId.value)
-  if (!startPos) {
-    connectionLines.value = []
-    return
-  }
-
-  const sameWeapons = getSameStatWeapons(hoveredWeaponId.value)
-  const newLines: ConnectionLine[] = []
-
-  for (const targetId of sameWeapons) {
-    const endPos = getWeaponElementPosition(targetId)
-    if (!endPos) continue
-
-    const { length, angle } = calculateLine(
-      startPos.x,
-      startPos.y,
-      endPos.x,
-      endPos.y,
-    )
-
-    newLines.push({
-      style: {
-        left: `${startPos.x}px`,
-        top: `${startPos.y}px`,
-        width: `${length}px`,
-        transform: `rotate(${angle}deg)`,
-        opacity: 1,
-      },
-    })
-  }
-
-  connectionLines.value = newLines
-}
-
 /** 鼠标进入武器 */
 function handleWeaponMouseEnter(weaponId: string) {
-  hoveredWeaponId.value = weaponId
-  updateConnectionLines()
-}
-
-/** 鼠标离开武器 */
-function handleWeaponMouseLeave() {
-  hoveredWeaponId.value = null
-  connectionLines.value = []
+  handleWeaponMouseEnterBase(weaponId, getSameStatWeapons(weaponId))
 }
 
 // 监听窗口大小变化，更新连线位置
 onMounted(() => {
-  window.addEventListener('resize', updateConnectionLines)
-})
-
-onUnmounted(() => {
-  window.removeEventListener('resize', updateConnectionLines)
-  connectionLines.value = []
-  // 清理防抖定时器，避免组件销毁后发起无效请求
-  if (detailSaveTimer) {
-    clearTimeout(detailSaveTimer)
-    detailSaveTimer = null
-  }
-})
-
-// --- 自定义基质相关 ---
-
-/** 自定义宝藏基质属性配置列表 */
-const customStats = ref<Array<{ name: string; attribute: string | null; secondary: string | null; skill: string | null; no_prompt_switch?: boolean }>>([])
-
-// --- 详情弹窗编辑状态 ---
-
-/** 新建模式标识 */
-const isNewCustom = computed(() => detailWeaponId.value === '__new_custom__')
-
-/** 自定义基质编辑中的属性 */
-const customEditAttribute = ref<string | null>(null)
-const customEditSecondary = ref<string | null>(null)
-const customEditSkill = ref<string | null>(null)
-
-/** 等级编辑状态 */
-const detailAffix1 = ref(1)
-const detailAffix2 = ref(1)
-const detailAffix3 = ref(1)
-
-/** 优先级编辑状态 */
-const detailPriority = ref(6)
-
-/** 新建模式下的拥有状态（用户手动切换） */
-const detailOwnedOverride = ref(false)
-
-/** 当前弹窗中的条目是否已拥有 */
-const isDetailOwned = computed(() => {
-  const weaponId = detailWeaponId.value
-  if (!weaponId) return false
-  if (weaponId === '__new_custom__') return detailOwnedOverride.value
-  return isWeaponOwned(weaponId)
-})
-
-/** 切换当前弹窗条目的拥有状态 */
-async function toggleDetailOwnership() {
-  const weaponId = detailWeaponId.value
-  if (!weaponId) return
-  if (weaponId === '__new_custom__') {
-    detailOwnedOverride.value = !detailOwnedOverride.value
-  } else {
-    await toggleWeaponOwnership(weaponId)
-  }
-}
-
-// --- 自定义基质与内置武器重合检测 ---
-
-interface OverlapItem {
-  customIndex: number
-  customName: string
-  customWeaponId: string
-  matchedWeaponId: string
-  matchedWeaponName: string
-  action: 'ignore' | 'suppress' | 'switch' | 'delete'
-}
-
-const overlapItems = ref<OverlapItem[]>([])
-const overlapDialog = ref(false)
-/** 重合检测的等级比较结果（按 overlapItems 索引缓存） */
-const overlapCompareResults = ref<Map<number, number>>(new Map())
-
-/** 从后端获取配置中的自定义宝藏基质属性列表 */
-async function fetchCustomStats() {
-  try {
-    const res = await fetch('/api/config')
-    if (!res.ok) {
-      throw new Error(`HTTP ${res.status}: ${res.statusText}`)
-    }
-    const config = await res.json()
-    customStats.value = config.treasure_essence_stats || []
-  } catch (error) {
-    console.error('获取自定义宝藏基质配置失败:', error)
-  }
-}
-
-/** 判断是否为自定义基质条目（weapon_id 以 custom_stat_ 开头） */
-function isCustomEntry(weaponId: string | null): boolean {
-  return weaponId?.startsWith('custom_stat_') ?? false
-}
-
-/** 自定义基质条目列表，用于武器总览展示 */
-const customMatrixEntries = computed(() => {
-  return customStats.value
-    .map((stat, index) => ({
-      syntheticId: `custom_stat_${index}`,
-      displayName: stat.name || `自定义基质 ${index + 1}`,
-      index,
-      skillStatId: stat.skill,
-    }))
+  setupResizeListener(() => {
+    if (!hoveredWeaponId.value) return []
+    return getSameStatWeapons(hoveredWeaponId.value)
+  })
 })
 
 /** 是否显示自定义基质区段（当 6★ 筛选激活时显示） */
 const showCustomSection = computed(() => selectedRarities.value.includes('custom'))
 
-/** 自定义条目编辑中的名称 */
-const customEntryName = ref('')
-
-// --- 删除自定义基质 ---
-const deleteCustomConfirm = ref(false)
-const deleteCustomIndex = ref<number | null>(null)
-const deleteCustomName = ref('')
-// 标记「正在发起删除」：删除按钮按下（mousedown）时置位，用于让名称输入框的
-// @blur 保存跳过本次写入，避免与删除流程交错写后端。
-const pendingCustomDelete = ref(false)
-
-// 弹窗打开时，加载编辑状态
-watch([detailDialog, detailWeaponId], () => {
-  // 弹窗关闭时复位删除标记，避免 mousedown 后放弃点击导致标记卡住
-  if (!detailDialog.value) {
-    pendingCustomDelete.value = false
-    return
-  }
-
-  if (detailDialog.value) {
-    const weaponId = detailWeaponId.value
-
-    if (weaponId === '__new_custom__') {
-      // 新建模式：重置所有字段
-      customEntryName.value = ''
-      customEditAttribute.value = null
-      customEditSecondary.value = null
-      customEditSkill.value = null
-      detailAffix1.value = 1
-      detailAffix2.value = 1
-      detailAffix3.value = 1
-      detailPriority.value = 6
-      detailOwnedOverride.value = false
-    } else if (isCustomEntry(weaponId)) {
-      // 编辑自定义条目
-      const index = Number.parseInt(weaponId!.replace('custom_stat_', ''), 10)
-      const stat = customStats.value[index]
-      customEntryName.value = stat?.name || ''
-      customEditAttribute.value = stat?.attribute ?? null
-      customEditSecondary.value = stat?.secondary ?? null
-      customEditSkill.value = stat?.skill ?? null
-      // 从 profile 中读取等级
-      const entry = matrixEntryByWeaponId.value.get(weaponId!)
-      detailAffix1.value = entry?.affix1_level ?? 1
-      detailAffix2.value = entry?.affix2_level ?? 1
-      detailAffix3.value = entry?.affix3_level ?? 1
-      detailPriority.value = getWeaponPriority(weaponId!)
-    } else {
-      // 非自定义条目：从 profile 中读取等级
-      const entry = matrixEntryByWeaponId.value.get(weaponId!)
-      detailAffix1.value = entry?.affix1_level ?? 1
-      detailAffix2.value = entry?.affix2_level ?? 1
-      detailAffix3.value = entry?.affix3_level ?? 1
-      detailPriority.value = getWeaponPriority(weaponId!)
-    }
-  }
-})
-
-// 确认弹窗以任意方式关闭（取消/遮罩/ESC）时复位删除标记，避免误跳过后续名称保存
-watch(deleteCustomConfirm, (open) => {
-  if (!open) pendingCustomDelete.value = false
-})
-
 /** 打开新建自定义基质弹窗 */
 function showNewCustomDialog() {
   detailWeaponId.value = '__new_custom__'
-  detailDialog.value = true
 }
 
-/** 保存自定义基质（新建或编辑） */
-async function saveCustomEntry() {
-  const weaponId = detailWeaponId.value
-  if (!weaponId) return
-
-  try {
-    if (weaponId === '__new_custom__') {
-      // 新建模式：准备新条目数据
-      const newEntry = {
-        name: customEntryName.value || `自定义基质 ${customStats.value.length + 1}`,
-        attribute: customEditAttribute.value,
-        secondary: customEditSecondary.value,
-        skill: customEditSkill.value,
-      }
-
-      // 先发送请求，成功后再更新本地状态
-      const tempStats = [...customStats.value, newEntry]
-      await postCustomStatsUpdate(tempStats)
-
-      // 请求成功，更新本地状态
-      customStats.value.push(newEntry)
-
-      // 只有标记为"已拥有"时才写入 profile
-      if (isDetailOwned.value) {
-        const newIndex = customStats.value.length - 1
-        const syntheticId = `custom_stat_${newIndex}`
-        await addTreasureMatrixEntry({
-          weapon_id: syntheticId,
-          weapon_name: customEntryName.value || `自定义基质 ${newIndex + 1}`,
-          affix1_level: detailAffix1.value,
-          affix2_level: detailAffix2.value,
-          affix3_level: detailAffix3.value,
-          priority: detailPriority.value,
-          include_in_calculation: true,
-        })
-      }
-    } else if (isCustomEntry(weaponId)) {
-      // 编辑模式：准备更新数据
-      const index = Number.parseInt(weaponId.replace('custom_stat_', ''), 10)
-      if (customStats.value[index]) {
-        const originalEntry = { ...customStats.value[index] }
-        const updatedEntry = {
-          ...originalEntry,
-          name: customEntryName.value,
-          attribute: customEditAttribute.value,
-          secondary: customEditSecondary.value,
-          skill: customEditSkill.value,
-        }
-
-        // 先发送请求，成功后再更新本地状态
-        const tempStats = [...customStats.value]
-        tempStats[index] = updatedEntry
-        await postCustomStatsUpdate(tempStats)
-
-        // 请求成功，更新本地状态
-        customStats.value[index] = updatedEntry
-      }
-
-      // 更新 profile 中的等级和优先级
-      const entry = matrixEntryByWeaponId.value.get(weaponId)
-      if (entry) {
-        const originalEntry = { ...entry }
-        entry.weapon_name = customEntryName.value
-        entry.affix1_level = detailAffix1.value
-        entry.affix2_level = detailAffix2.value
-        entry.affix3_level = detailAffix3.value
-        entry.priority = detailPriority.value
-
-        try {
-          await updateTreasureMatrix([...treasureMatrix.value])
-          await updateWeaponPriority(weaponId, detailPriority.value)
-        } catch (profileError) {
-          // 回滚本地状态
-          Object.assign(entry, originalEntry)
-          throw profileError
-        }
-      }
-    }
-
-    detailDialog.value = false
-    await fetchCustomStats()
-  } catch (error) {
-    console.error('保存自定义基质失败:', error)
-    // TODO: 显示 snackbar 错误提示
-  }
-}
-
-/** 非自定义基质：从基质配置中移除 */
-async function removeNonCustomEntry(weaponId: string) {
-  try {
-    await removeTreasureMatrixEntry(weaponId)
-    detailDialog.value = false
-  } catch (error) {
-    console.error('移除基质条目失败:', error)
-    // TODO: 显示 snackbar 错误提示
-  }
-}
-
-/** 非自定义基质：等级/优先级变化时自动保存（防抖） */
-let detailSaveTimer: ReturnType<typeof setTimeout> | null = null
-watch([detailAffix1, detailAffix2, detailAffix3, detailPriority], async () => {
-  const weaponId = detailWeaponId.value
-  // 仅对已存在的非自定义条目自动保存
-  if (!weaponId || weaponId === '__new_custom__' || isCustomEntry(weaponId)) return
-  if (!isWeaponOwned(weaponId)) return
-
-  if (detailSaveTimer) clearTimeout(detailSaveTimer)
-  detailSaveTimer = setTimeout(async () => {
-    try {
-      const entry = matrixEntryByWeaponId.value.get(weaponId)
-      if (entry) {
-        entry.affix1_level = detailAffix1.value
-        entry.affix2_level = detailAffix2.value
-        entry.affix3_level = detailAffix3.value
-        entry.priority = detailPriority.value
-        await updateTreasureMatrix([...treasureMatrix.value])
-        await updateWeaponPriority(weaponId, detailPriority.value)
-      }
-    } catch (error) {
-      console.error('自动保存等级/优先级失败:', error)
-      // TODO: 显示 snackbar 错误提示
-    } finally {
-      detailSaveTimer = null
-    }
-  }, 400)
-})
-
-/** 打开删除自定义基质的二次确认弹窗 */
-function promptDeleteCustomEntry() {
-  if (!isCustomEntry(detailWeaponId.value)) return
-  const index = Number.parseInt(detailWeaponId.value!.replace('custom_stat_', ''), 10)
-  if (Number.isNaN(index)) return
-  deleteCustomIndex.value = index
-  deleteCustomName.value = customStats.value[index]?.name || `自定义基质 ${index + 1}`
-  deleteCustomConfirm.value = true
-}
-
-/** 取消删除 */
-function cancelDeleteCustomEntry() {
-  deleteCustomConfirm.value = false
-  pendingCustomDelete.value = false
-}
-
-/**
- * 确认删除自定义基质：
- * 1. 基于未修改的 treasureMatrix 计算新矩阵（删除目标条目 + 重索引后续自定义条目）
- * 2. 从 config 的自定义列表移除该项
- * 3. 先写 profile（后端据此重建 weapon_priorities），再回写 config，最后回读
- */
-async function confirmDeleteCustomEntry() {
-  const index = deleteCustomIndex.value
-  if (index === null) {
-    cancelDeleteCustomEntry()
-    return
-  }
-  const targetId = `custom_stat_${index}`
-
-  try {
-    // 1) 删除目标条目，并将所有索引 > index 的自定义条目前移一位
-    const newMatrix = treasureMatrix.value
-      .filter((e) => e.weapon_id !== targetId)
-      .map((e) => {
-        if (e.weapon_id.startsWith('custom_stat_')) {
-          const cur = Number.parseInt(e.weapon_id.replace('custom_stat_', ''), 10)
-          if (cur > index) return { ...e, weapon_id: `custom_stat_${cur - 1}` }
-        }
-        return e
-      })
-
-    // 2) 准备新的 config 数据（不修改本地状态）
-    const tempStats = customStats.value.filter((_, i) => i !== index)
-
-    // 3) 持久化：先 profile，再 config，最后回读配置
-    await updateTreasureMatrix(newMatrix)
-    await postCustomStatsUpdate(tempStats)
-
-    // 4) 请求成功，更新本地状态
-    customStats.value.splice(index, 1)
-    await fetchCustomStats()
-
-    // 5) 关闭弹窗并清理引用（重索引后旧 syntheticId 已指向不同条目，必须清空）
-    deleteCustomConfirm.value = false
-    pendingCustomDelete.value = false
-    detailDialog.value = false
-    detailWeaponId.value = null
-    deleteCustomIndex.value = null
-  } catch (error) {
-    console.error('删除自定义基质失败:', error)
-    // TODO: 显示 snackbar 错误提示
-  }
-}
-
-/** 将自定义宝藏基质配置保存到后端 */
-async function postCustomStatsUpdate(stats?: CustomStat[]) {
-  const getRes = await fetch('/api/config')
-  if (!getRes.ok) {
-    throw new Error(`HTTP ${getRes.status}: ${getRes.statusText}`)
-  }
-  const currentConfig = await getRes.json()
-  currentConfig.treasure_essence_stats = stats ?? customStats.value
-  const postRes = await fetch('/api/config', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(currentConfig),
-  })
-  if (!postRes.ok) {
-    throw new Error(`HTTP ${postRes.status}: ${postRes.statusText}`)
-  }
-}
-
-/** 获取重合检测中某个武器 ID 的属性词条文本（含等级） */
-function getOverlapTooltipText(weaponId: string): string {
-  const entry = matrixEntryByWeaponId.value.get(weaponId)
-  const levels: [number, number, number] = [
-    entry?.affix1_level ?? 0,
-    entry?.affix2_level ?? 0,
-    entry?.affix3_level ?? 0,
-  ]
-
-  let statIds: (string | null)[] = []
-  if (isCustomEntry(weaponId)) {
-    const index = Number.parseInt(weaponId.replace('custom_stat_', ''), 10)
-    const stat = customStats.value[index]
-    statIds = [stat?.attribute ?? null, stat?.secondary ?? null, stat?.skill ?? null]
-  } else {
-    const weapon = weaponsMap.value.get(weaponId)
-    statIds = [weapon?.attributeStatId ?? null, weapon?.secondaryStatId ?? null, weapon?.skillStatId ?? null]
-  }
-
-  const parts: string[] = []
-  for (let i = 0; i < 3; i++) {
-    const sid = statIds[i]
-    if (sid) {
-      parts.push(`${getGemTagName(sid)} Lv.${levels[i]}`)
-    }
-  }
-  return parts.join('、') || '无属性'
-}
-
-/**
- * 解析武器 ID 对应的三个槽位词条类型
- * 返回 ["ATTRIBUTE", "SECONDARY", "SKILL"] 或对应的 null
- */
-function resolveStatTypes(weaponId: string): (string | null)[] {
-  if (isCustomEntry(weaponId)) {
-    const index = Number.parseInt(weaponId.replace('custom_stat_', ''), 10)
-    const stat = customStats.value[index]
-    return [
-      stat?.attribute ? essencesMap.value.get(stat.attribute)?.type ?? null : null,
-      stat?.secondary ? essencesMap.value.get(stat.secondary)?.type ?? null : null,
-      stat?.skill ? essencesMap.value.get(stat.skill)?.type ?? null : null,
-    ]
-  }
-  const weapon = weaponsMap.value.get(weaponId)
-  return [
-    weapon?.attributeStatId ? 'ATTRIBUTE' : null,
-    weapon?.secondaryStatId ? 'SECONDARY' : null,
-    weapon?.skillStatId ? 'SKILL' : null,
-  ]
-}
-
-/** 从后端批量获取重合检测的等级比较结果 */
-async function fetchOverlapCompareResults() {
-  if (overlapItems.value.length === 0) return
-  try {
-    const results = new Map<number, number>()
-    const res = await fetch('/api/profiles/compare_levels', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({
-        items: overlapItems.value.map((item) => {
-          const customEntry = matrixEntryByWeaponId.value.get(item.customWeaponId)
-          const matchedEntry = matrixEntryByWeaponId.value.get(item.matchedWeaponId)
-          return {
-            current_levels: [
-              customEntry?.affix1_level ?? 0,
-              customEntry?.affix2_level ?? 0,
-              customEntry?.affix3_level ?? 0,
-            ],
-            existing_levels: [
-              matchedEntry?.affix1_level ?? 0,
-              matchedEntry?.affix2_level ?? 0,
-              matchedEntry?.affix3_level ?? 0,
-            ],
-            stat_types: resolveStatTypes(item.customWeaponId),
-          }
-        }),
-      }),
-    })
-    if (!res.ok) {
-      throw new Error(`HTTP ${res.status}: ${res.statusText}`)
-    }
-    const data = await res.json()
-    for (let i = 0; i < overlapItems.value.length; i++) {
-      results.set(i, data.results?.[i] ?? 0)
-    }
-    overlapCompareResults.value = results
-  } catch (error) {
-    console.error('获取等级比较结果失败:', error)
-  }
-}
-
-/** 获取预计算的比较结果：1（自定义更优）/ 0（相等）/ -1（内置更优） */
-function getOverlapCompareResult(item: OverlapItem): number {
-  const idx = overlapItems.value.indexOf(item)
-  return overlapCompareResults.value.get(idx) ?? 0
-}
-
-/** 一键选择：根据比较结果自动选择动作 */
-function autoSelectOverlap() {
-  for (let i = 0; i < overlapItems.value.length; i++) {
-    const item = overlapItems.value[i]
-    if (!item) continue
-    const cmp = overlapCompareResults.value.get(i) ?? 0
-    if (cmp < 0) {
-      item.action = 'delete'   // 自定义更小 → 删除自定义
-    } else if (cmp === 0) {
-      item.action = 'ignore'   // 相等 → 忽略
-    } else {
-      item.action = 'switch'   // 自定义更大 → 切换到内置
-    }
-  }
-}
-
-/**
- * 检查自定义基质与内置武器的词条重合
- * 匹配规则：三个槽位完全相等（含 null 对 null）
- * 检查范围：所有 config 中配置的自定义基质，无论是否在 profiles 中拥有
- */
+/** 检查自定义基质重合 */
 function checkCustomOverlap() {
-  const items: OverlapItem[] = []
-  for (let i = 0; i < customStats.value.length; i++) {
-    const stat = customStats.value[i]
-    if (!stat) continue
-    // 跳过已勾选"不再提示"的
-    if (stat.no_prompt_switch) continue
-    // 跳过属性全为空的条目（已被切换清空）
-    if (!stat.attribute && !stat.secondary && !stat.skill) continue
-
-    const syntheticId = `custom_stat_${i}`
-
-    // 遍历所有内置武器，查找三词条完全匹配
-    for (const [weaponId, weapon] of weaponsMap.value.entries()) {
-      if (
-        weapon.attributeStatId === stat.attribute &&
-        weapon.secondaryStatId === stat.secondary &&
-        weapon.skillStatId === stat.skill
-      ) {
-        items.push({
-          customIndex: i,
-          customName: stat.name || `自定义基质 ${i + 1}`,
-          customWeaponId: syntheticId,
-          matchedWeaponId: weaponId,
-          matchedWeaponName: weapon.name,
-          action: 'ignore',
-        })
-      }
-    }
-  }
-  if (items.length === 0) return
-  overlapItems.value = items
-  overlapDialog.value = true
-  fetchOverlapCompareResults()
-}
-
-/** 确认重合操作 */
-async function confirmOverlapActions() {
-  // 记录需要删除的自定义基质索引
-  const indicesToDelete: number[] = []
-  // 记录是否有 suppress 操作需要更新配置
-  let hasSuppress = false
-
-  // 第一轮：收集所有修改，不修改本地状态
-  for (const item of overlapItems.value) {
-    if (item.action === 'ignore') continue
-
-    const stat = customStats.value[item.customIndex]
-    if (!stat) continue
-
-    if (item.action === 'suppress') {
-      hasSuppress = true
-    }
-
-    if (item.action === 'switch' || item.action === 'delete') {
-      indicesToDelete.push(item.customIndex)
-    }
-  }
-
-  // 从大到小排序删除，避免索引偏移问题
-  indicesToDelete.sort((a, b) => b - a)
-
-  // 构建新的 treasure_matrix（一次性完成所有 switch 和 delete 操作）
-  const toRemove = new Set(
-    overlapItems.value
-      .filter((i) => i.action === 'switch' || i.action === 'delete')
-      .map((i) => i.customWeaponId),
-  )
-  let newTreasureMatrix = treasureMatrix.value.filter((e) => !toRemove.has(e.weapon_id))
-
-  // 处理 switch 操作：追加内置武器条目
-  for (const item of overlapItems.value) {
-    if (item.action !== 'switch') continue
-
-    const entry = matrixEntryByWeaponId.value.get(item.customWeaponId)
-    if (entry) {
-      const weapon = weaponsMap.value.get(item.matchedWeaponId)
-      newTreasureMatrix.push({
-        weapon_id: item.matchedWeaponId,
-        weapon_name: weapon?.name || item.matchedWeaponName,
-        affix1_level: entry.affix1_level,
-        affix2_level: entry.affix2_level,
-        affix3_level: entry.affix3_level,
-        include_in_calculation: entry.include_in_calculation,
-        // priority 不设置，使用武器稀有度默认值
-      })
-    }
-  }
-
-  // 准备新的 config 数据（不修改本地状态）
-  let tempStats = [...customStats.value]
-  for (const index of indicesToDelete) {
-    tempStats = tempStats.filter((_, i) => i !== index)
-  }
-
-  // 更新 suppress 标记
-  if (hasSuppress) {
-    for (const item of overlapItems.value) {
-      if (item.action === 'suppress' && tempStats[item.customIndex]) {
-        const stat = tempStats[item.customIndex]!
-        tempStats[item.customIndex] = { ...stat, no_prompt_switch: true }
-      }
-    }
-  }
-
-  // 更新 treasure_matrix 中所有引用后续自定义基质的索引
-  if (indicesToDelete.length > 0) {
-    newTreasureMatrix = newTreasureMatrix.map((e) => {
-      if (e.weapon_id.startsWith('custom_stat_')) {
-        const currentIndex = Number.parseInt(e.weapon_id.replace('custom_stat_', ''), 10)
-        // 计算删除后的新索引
-        let newIndex = currentIndex
-        for (const deletedIndex of indicesToDelete) {
-          if (currentIndex > deletedIndex) {
-            newIndex--
-          }
-        }
-        if (newIndex !== currentIndex) {
-          return { ...e, weapon_id: `custom_stat_${newIndex}` }
-        }
-      }
-      return e
-    })
-  }
-
-  // 第二轮：一次性提交所有修改到后端
-  try {
-    if (indicesToDelete.length > 0) {
-      await updateTreasureMatrix(newTreasureMatrix)
-    }
-    if (hasSuppress || indicesToDelete.length > 0) {
-      await postCustomStatsUpdate(tempStats)
-    }
-
-    // 请求成功，更新本地状态
-    for (const index of indicesToDelete) {
-      customStats.value.splice(index, 1)
-    }
-    if (hasSuppress) {
-      for (const item of overlapItems.value) {
-        if (item.action === 'suppress' && customStats.value[item.customIndex]) {
-          customStats.value[item.customIndex]!.no_prompt_switch = true
-        }
-      }
-    }
-  } catch (error) {
-    console.error('确认重合操作失败:', error)
-    // TODO: 显示 snackbar 错误提示
-  }
-
-  overlapDialog.value = false
-  await fetchCustomStats()
+  checkCustomOverlapBase(customStats.value, matrixEntryByWeaponId.value)
 }
 
 onMounted(async () => {
   await fetchCustomStats()
   checkCustomOverlap()
 })
-
-const matrixEntryByWeaponId = computed(
-  () => new Map(treasureMatrix.value.map((entry) => [entry.weapon_id, entry])),
-)
-
-// 武器总览会为每个图标多次判断拥有/满级状态，用 Set/Map 避免重复扫描 treasureMatrix。
-const ownedWeaponIds = computed(() => new Set(matrixEntryByWeaponId.value.keys()))
 
 const totalCount = computed(() =>
   weaponTypes.value.reduce((sum, wType) => sum + wType.weaponIds.length, 0),
@@ -1503,50 +430,11 @@ const filteredWeaponTypes = computed(() => {
     .filter((wType) => wType.weaponIds.length > 0)
 })
 
-function isWeaponOwned(weaponId: string): boolean {
-  return ownedWeaponIds.value.has(weaponId)
-}
-
-function isWeaponMaxed(weaponId: string): boolean {
-  const entry = matrixEntryByWeaponId.value.get(weaponId)
-  return (
-    entry !== undefined &&
-    entry.affix1_level === 6 &&
-    entry.affix2_level === 6 &&
-    entry.affix3_level === 3
-  )
-}
-
-async function toggleWeaponOwnership(weaponId: string) {
-  if (isWeaponOwned(weaponId)) {
-    await removeTreasureMatrixEntry(weaponId)
-  } else {
-    // 自定义条目使用配置中的名称，普通武器使用 weaponsMap 中的名称
-    let weaponName: string
-    if (isCustomEntry(weaponId)) {
-      const index = Number.parseInt(weaponId.replace('custom_stat_', ''), 10)
-      weaponName = customStats.value[index]?.name || `自定义基质 ${index + 1}`
-    } else {
-      const weapon = weaponsMap.value.get(weaponId)
-      weaponName = weapon?.name || weaponId
-    }
-    await addTreasureMatrixEntry({
-      weapon_id: weaponId,
-      weapon_name: weaponName,
-      affix1_level: 1,
-      affix2_level: 1,
-      affix3_level: 1,
-      include_in_calculation: true,
-    })
-  }
-}
-
 /**
  * 显示武器详情弹窗（左键点击）
  */
 function showWeaponDetail(weaponId: string) {
   detailWeaponId.value = weaponId
-  detailDialog.value = true
 }
 
 /**
@@ -1581,166 +469,6 @@ function getWeaponStatsText(weaponId: string): string {
   }
 
   return parts.join('、') || '无属性'
-}
-
-/**
- * 获取同类武器（相同属性组合）
- * 自定义条目没有真实武器数据，返回空数组
- */
-function getSameStatWeapons(weaponId: string): string[] {
-  if (isCustomEntry(weaponId)) return []
-  const weapon = weaponsMap.value.get(weaponId)
-  if (!weapon) return []
-  const sameWeapons: string[] = []
-  for (const [id, w] of weaponsMap.value.entries()) {
-    if (
-      id !== weaponId &&
-      w.attributeStatId === weapon.attributeStatId &&
-      w.secondaryStatId === weapon.secondaryStatId &&
-      w.skillStatId === weapon.skillStatId
-    ) {
-      sameWeapons.push(id)
-    }
-  }
-  return sameWeapons
-}
-
-/**
- * 判断武器是否"可切换"：存在同属性、更高优先级、且已拥有的武器
- */
-function isSwitchable(weaponId: string): boolean {
-  const myPriority = getWeaponPriority(weaponId)
-  const sameWeapons = getSameStatWeapons(weaponId)
-  return sameWeapons.some(
-    (id) => isWeaponOwned(id) && getWeaponPriority(id) >= myPriority,
-  )
-}
-
-/**
- * 获取可切换的目标武器是否满级（用于灰色呼吸动画）
- */
-function isSwitchTargetMaxed(weaponId: string): boolean {
-  const myPriority = getWeaponPriority(weaponId)
-  const sameWeapons = getSameStatWeapons(weaponId)
-  return sameWeapons.some(
-    (id) =>
-      isWeaponOwned(id)
-      && getWeaponPriority(id) >= myPriority
-      && isWeaponMaxed(id),
-  )
-}
-
-/**
- * 获取武器的基质等级文本
- */
-function getMatrixLevelText(weaponId: string): string {
-  const entry = matrixEntryByWeaponId.value.get(weaponId)
-  if (!entry) return '未配置'
-  return `+${entry.affix1_level} / +${entry.affix2_level} / +${entry.affix3_level}`
-}
-
-/**
- * 获取用户手动设置的优先级（0 表示未设置）
- */
-function getUserPriority(weaponId: string): number {
-  const profilePriority = activeProfile.value.weapon_priorities?.[weaponId]
-  if (profilePriority && profilePriority > 0) return profilePriority
-  const entry = matrixEntryByWeaponId.value.get(weaponId)
-  return entry?.priority || 0
-}
-
-/**
- * 获取武器的有效优先级（未设置时使用稀有度）
- * 自定义条目默认优先级为 6（等同于 6★）
- */
-function getWeaponPriority(weaponId: string): number {
-  const userP = getUserPriority(weaponId)
-  if (userP > 0) return userP
-  // 自定义条目默认优先级为 6
-  if (isCustomEntry(weaponId)) return 6
-  const weapon = weaponsMap.value.get(weaponId)
-  return weapon ? weapon.rarity : 0
-}
-
-function getEffectivePriorityForSwap(weaponId: string, entry?: TreasureMatrixEntry): number {
-  const userPriority = getUserPriority(weaponId) || entry?.priority || 0
-  if (userPriority > 0) return userPriority
-  const weapon = weaponsMap.value.get(weaponId)
-  return weapon ? weapon.rarity : 0
-}
-
-/**
- * 交换两把武器的基质数据
- */
-async function swapMatrix(weaponAId: string, weaponBId: string) {
-  const entries = treasureMatrix.value.map((entry) => ({ ...entry }))
-  const entryA = entries.find((e) => e.weapon_id === weaponAId)
-  const entryB = entries.find((e) => e.weapon_id === weaponBId)
-
-  const weaponA = weaponsMap.value.get(weaponAId)
-  const weaponB = weaponsMap.value.get(weaponBId)
-
-  const hasA = !!entryA
-  const hasB = !!entryB
-
-  if (!hasA && !hasB) return
-
-  const priorityA = getEffectivePriorityForSwap(weaponAId, entryA)
-  const priorityB = getEffectivePriorityForSwap(weaponBId, entryB)
-
-  if (hasA && !hasB) {
-    // A有基质、B无基质 → A移除、B添加A的数据
-    const nextEntries = entries
-      .filter((entry) => entry.weapon_id !== weaponAId)
-      .concat({
-        ...entryA!,
-        weapon_id: weaponBId,
-        weapon_name: weaponB?.name || weaponBId,
-        priority: priorityA,
-      })
-    await updateTreasureMatrix(nextEntries)
-    await updateWeaponPriority(weaponAId, priorityB)
-    await updateWeaponPriority(weaponBId, priorityA)
-  } else if (!hasA && hasB) {
-    // A无基质、B有基质 → A添加B的数据、B移除
-    const nextEntries = entries
-      .filter((entry) => entry.weapon_id !== weaponBId)
-      .concat({
-        ...entryB!,
-        weapon_id: weaponAId,
-        weapon_name: weaponA?.name || weaponAId,
-        priority: priorityB,
-      })
-    await updateTreasureMatrix(nextEntries)
-    await updateWeaponPriority(weaponAId, priorityB)
-    await updateWeaponPriority(weaponBId, priorityA)
-  } else {
-    // 两者都有基质，交换等级、计算开关和有效优先级
-    const matrixA = {
-      affix1: entryA!.affix1_level,
-      affix2: entryA!.affix2_level,
-      affix3: entryA!.affix3_level,
-      includeInCalculation: entryA!.include_in_calculation,
-    }
-    entryA!.affix1_level = entryB!.affix1_level
-    entryA!.affix2_level = entryB!.affix2_level
-    entryA!.affix3_level = entryB!.affix3_level
-    entryA!.include_in_calculation = entryB!.include_in_calculation
-    entryA!.priority = priorityB
-
-    entryB!.affix1_level = matrixA.affix1
-    entryB!.affix2_level = matrixA.affix2
-    entryB!.affix3_level = matrixA.affix3
-    entryB!.include_in_calculation = matrixA.includeInCalculation
-    entryB!.priority = priorityA
-
-    await updateTreasureMatrix(entries)
-    await updateWeaponPriority(weaponAId, priorityB)
-    await updateWeaponPriority(weaponBId, priorityA)
-  }
-
-  // 关闭弹窗
-  detailDialog.value = false
 }
 </script>
 
@@ -1811,99 +539,6 @@ async function swapMatrix(weaponAId: string, weaponBId: string) {
   &:hover {
     border-color: rgba(var(--v-theme-primary), 0.6);
     background: rgba(var(--v-theme-primary), 0.05);
-  }
-}
-
-// --- 详情弹窗等级编辑样式 ---
-.detail-level-section {
-  display: flex;
-  gap: 16px;
-  flex-wrap: wrap;
-}
-
-.detail-attr-control {
-  flex: 1;
-  min-width: 120px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 6px;
-  padding: 10px 12px;
-  border: 1px solid rgba(var(--v-border-color), 0.12);
-  border-radius: 12px;
-  transition: background 0.18s;
-
-  &:hover {
-    background: rgba(var(--v-theme-on-surface), 0.03);
-  }
-}
-
-.detail-attr-label {
-  color: rgba(var(--v-theme-on-surface), 0.52);
-  font-size: 0.72rem;
-  font-weight: 700;
-  letter-spacing: 0.06em;
-  text-transform: uppercase;
-}
-
-.detail-attr-pips {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  min-height: 20px;
-}
-
-.detail-pip {
-  width: 10px;
-  height: 20px;
-  border-radius: 999px;
-  background: rgba(var(--v-theme-on-surface), 0.12);
-  cursor: pointer;
-  transition: background 0.18s, box-shadow 0.18s, transform 0.18s;
-
-  &:hover {
-    transform: translateY(-1px) scaleY(1.08);
-  }
-}
-
-.detail-attr-control--primary .detail-pip.active {
-  background: rgb(var(--v-theme-primary));
-  box-shadow: 0 2px 7px rgba(var(--v-theme-primary), 0.32);
-}
-
-.detail-attr-control--teal .detail-pip.active {
-  background: #48a9a6;
-  box-shadow: 0 2px 7px rgba(72, 169, 166, 0.34);
-}
-
-.detail-attr-control--indigo .detail-pip.active {
-  background: #5c6bc0;
-  box-shadow: 0 2px 7px rgba(92, 107, 192, 0.34);
-}
-
-.detail-pip.detail-pip--max.active {
-  animation: detailPipPulse 2.2s ease-in-out infinite;
-}
-
-.detail-attr-value {
-  color: rgba(var(--v-theme-on-surface), 0.68);
-  font-family: 'JetBrains Mono', 'Fira Code', monospace;
-  font-size: 0.85rem;
-  font-weight: 800;
-}
-
-.detail-attr-value--full {
-  color: rgb(var(--v-theme-primary));
-}
-
-@keyframes detailPipPulse {
-  0%, 100% {
-    transform: scaleY(1);
-    filter: brightness(1);
-  }
-  50% {
-    transform: scaleY(1.08);
-    filter: brightness(1.18);
   }
 }
 
@@ -1999,110 +634,6 @@ async function swapMatrix(weaponAId: string, weaponBId: string) {
   z-index: 5;
   pointer-events: none;
   box-shadow: 0 0 4px rgba(255, 113, 0, 0.6);
-}
-
-// 未拥有斜向胶带遮罩（弹窗内等级区域）
-.detail-level-wrapper {
-  position: relative;
-}
-
-.not-owned-tape-detail {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  width: 110%;
-  height: 24px;
-  background: linear-gradient(
-    135deg,
-    rgba(255, 193, 7, 0.95) 0%,
-    rgba(255, 193, 7, 0.88) 40%,
-    rgba(255, 193, 7, 0.80) 100%
-  );
-  transform: translate(-50%, -50%) rotate(-8deg);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 5;
-  pointer-events: none;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.25);
-}
-
-.not-owned-tape-detail-text {
-  color: rgba(0, 0, 0, 0.8);
-  font-size: 0.8rem;
-  font-weight: 900;
-  letter-spacing: 0.2em;
-  white-space: nowrap;
-}
-
-// 撕开胶布动画（左到右渐出/渐入）
-.tape-peel-enter-active {
-  animation: tape-peel-in 0.2s ease-out forwards;
-}
-
-.tape-peel-leave-active {
-  animation: tape-peel-out 0.2s ease-in forwards;
-}
-
-@keyframes tape-peel-in {
-  from {
-    clip-path: inset(0 100% 0 0);
-  }
-  to {
-    clip-path: inset(0 0 0 0);
-  }
-}
-
-@keyframes tape-peel-out {
-  from {
-    clip-path: inset(0 0 0 0);
-  }
-  to {
-    clip-path: inset(0 0 0 100%);
-  }
-}
-
-// 未拥有遮罩层：覆盖等级区域，阻止点击
-.detail-level-outer {
-  position: relative;
-}
-
-.not-owned-block-overlay {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  z-index: 6;
-  cursor: pointer;
-}
-
-.weapon-icon-detail {
-  width: 3rem !important;
-  height: 3rem !important;
-}
-
-.weapon-icon-same {
-  width: 2rem !important;
-  height: 2rem !important;
-  flex-shrink: 0;
-}
-
-.weapon-icon-overlap {
-  width: 2rem !important;
-  height: 2rem !important;
-  flex-shrink: 0;
-}
-
-.overlap-icon-wrapper {
-  width: 2rem;
-  height: 2rem;
-  flex-shrink: 0;
-
-  &.weapon-not-owned {
-    opacity: 0.4;
-    filter: grayscale(0.8);
-  }
 }
 
 .rainbow-border {
